feat(app): add local translation providers, model download, and test UI
- Apple Translation (macOS 26+ on-device) via Swift static library
- Marian/NLLB ONNX local model support with ort + tokenizers
- Model download flow from Hugging Face with progress events
- Translation test area in Settings for verifying provider config
- Source language selector (auto-detect + manual)
- Conditional OpenAI config panel (only shown when openai selected)
- Pin @tauri-apps/api and cli to 2.10.1 (fix minor version mismatch)
- Add .taurignore to prevent Swift .build watch loop in dev mode

+++ autocorrect-app/src-tauri/src/macos_text.rs
@@ -923,7 +923,7 @@ pub fn get_selected_text_bounds() -> Result<(i32, i32, i32, i32)> {
                         log::info!("[DIAG] AXBoundsForRange returned invalid bounds, using mouse position fallback");
                         let (mouse_x, mouse_y) = get_cursor_position_nsevent();
                         // 返回鼠标位置作为近似选区位置，使用固定的小宽度
-                        return Ok((mouse_x as i32, mouse_y as i32, 100, 20));
+                        return Ok((mouse_x, mouse_y, 100, 20));
                     }
                 }
             }
